Splitted the arbconverter.c file into multiple smaller ones
