+++ Android.mk
@@ -16,6 +16,9 @@ LOCAL_EXPORT_C_INCLUDES := $(LOCAL_C_INCLUDES) -DBCMHOST
 
 LOCAL_SRC_FILES := \
 	src/gl/arbconverter.c \
+	src/gl/arbgenerator.c \
+	src/gl/arbhelper.c \
+	src/gl/arbparser.c \
 	src/gl/array.c \
 	src/gl/blend.c \
 	src/gl/blit.c \
@@ -96,4 +99,3 @@ LOCAL_LDLIBS := -ldl -llog
 
 LOCAL_CFLAGS += -DSTATICLIB
 include $(BUILD_STATIC_LIBRARY)
-
