+++ src/CMakeLists.txt
@@ -3,6 +3,9 @@ include_directories(util)
 #file(GLOB_RECURSE GL_SOURCES gl/*.c)
 SET(GL_SRC
     gl/arbconverter.c
+	gl/arbgenerator.c
+	gl/arbhelper.c
+	gl/arbparser.c
     gl/array.c
     gl/blit.c
     gl/blend.c
@@ -71,6 +74,9 @@ SET(GL_SRC
 SET(GL_H
     ../version.h
     gl/arbconverter.h
+	gl/arbgenerator.h
+	gl/arbhelper.h
+	gl/arbparser.h
     gl/array.h
     gl/blend.h
     gl/blit.h
